Setup screen edit

+++ app/src/main/java/com/aayar94/aquatick/domain/model/SetupScreenModel.kt
@@ -3,7 +3,7 @@ package com.aayar94.aquatick.domain.model
 data class SetupScreenModel(
     val name: String? =null,
     val age: Int? = null,
-    val weight: Int? = null,
+    val weight: Double? = null,
     val gender: String? = null,
     val activityLevel: String? = null,
 )
