+++ app/src/main/java/com/aayar94/aquatick/ui/screen/setup/SetupViewModel.kt
@@ -54,7 +54,7 @@ class SetupViewModel @Inject constructor(
 
     fun updateWeight(weight: String) {
         _uiState.update {
-            it.copy(weight = weight.toInt())
+            it.copy(weight = weight.toDouble())
         }
     }
